+++ spec/rubocop/cop/naming/constant_name_spec.rb
@@ -63,12 +63,24 @@ RSpec.describe RuboCop::Cop::Naming::ConstantName do
     expect_no_offenses('AnythingGoes = test')
   end
 
+  it 'does not check names if rhs is a method call with conditional assign' do
+    expect_no_offenses('AnythingGoes ||= test')
+  end
+
   it 'does not check names if rhs is a `Class.new`' do
     expect_no_offenses('Invalid = Class.new(StandardError)')
   end
 
+  it 'does not check names if rhs is a `Class.new` with conditional assign' do
+    expect_no_offenses('Invalid ||= Class.new(StandardError)')
+  end
+
   it 'does not check names if rhs is a `Struct.new`' do
     expect_no_offenses('Investigation = Struct.new(:offenses, :errors)')
+  end
+
+  it 'does not check names if rhs is a `Struct.new` with conditional assign' do
+    expect_no_offenses('Investigation ||= Struct.new(:offenses, :errors)')
   end
 
   it 'does not check names if rhs is a method call with block' do